+++ CHANGELOG.md
@@ -19,5 +19,6 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Alphabetically sort `IsoCurrency`, `Holidays` and `BbanStructure` enums entries (#161).
 - Bump strata-basics from 2.12.5 to 2.12.6 (#165).
+- Add benchmarks that track the evolution of jbanking performances from version to version (#166).
 
 ### Thanks